Simplify HttpFileHandler.
Now, the HttpFileHandler just falls through if the requested file was not found. Allows executing other handlers to
handle that case, e.g. the new HttpSingleFileHandler that always serves the same file no matter which url was requested.
This is especially useful for React apps where all routes are handled by index.html.